API Reference

CommandComposer module

CommandComposer.Flag — Type

Flag(long_name, short_name, description)

Create a Flag object.

Arguments

long_name::String: the full name of the flag, e.g. "recursive" for a -r flag.

short_name::String: the short, usually single-letter, name of the flag, e.g. "r" for a --recursive flag.

description::String: a description of the flag for documentation purposes.

CommandComposer.Option — Type

Option(long_name, short_name, description, value)

Create an Option object.

Arguments

long_name::String: the full name of the option, e.g. "file" for a -f option.

short_name::String: the short, usually single-letter, name of the option, e.g. "f" for a --file option.

description::String: a description of the option for documentation purposes.

value::Any: the value assigned to this option.

CommandComposer.AndCommands — Type

AndCommands(a::AbstractCommand, b::AbstractCommand)

Represents a conjunction of two commands (i.e., both commands are executed).

CommandComposer.OrCommands — Type

OrCommands(a::AbstractCommand, b::AbstractCommand)

Represents a disjunction of two commands (i.e., either command is executed).

CommandComposer.RedirectedCommand — Type

RedirectedCommand(source, destination)

A RedirectedCommand is a wrapper around an AbstractCommand or a file (String).

It allows output redirection from a command to a file or input redirection from a file to a command. The redirection is specified by the RedirectedCommand's source and destination parameters.

Arguments

source: The source from which to redirect. If source is a String, it represents the filename of the file from which to read. If source is a subtype of AbstractCommand, it represents the command whose output is to be redirected.

destination: The destination to which to redirect. If destination is a String, it represents the filename of the file to which to write. If destination is a subtype of AbstractCommand, it represents the command that takes the redirected output as input.

CommandComposer.Command — Type

Command(name, flags, options, arguments, subcommands)

Represents a command to be executed, with associated flags, options, arguments, and subcommands.

Arguments

name::String: the name of the command.

flags::Vector{Flag}: any flags associated with the command.

options::Vector{Option}: any options associated with the command.

arguments::Vector{String}: any arguments to the command.

subcommands::Vector{AbstractCommand}: any subcommands to be executed in conjunction with the main command.

CommandComposer.interpret — Function

interpret(command::Command)

Translate a Command object into a Cmd that can be executed.

Examples

julia> c = Command("ls", [Flag("all", "a", "list all files")], [], [], []);

julia> interpret(c)
`ls --all`

interpret(command::RedirectedCommand)

Translate a RedirectedCommand object into a Base.CmdRedirect that can be executed, considering the redirection.

Examples

julia> r = RedirectedCommand(Command("ls", [], [], [], []), "output.txt");

julia> cmd = interpret(r)
pipeline(`ls`, stdout>Base.FileRedirect("output.txt", false))

julia> typeof(cmd)
Base.CmdRedirect

julia> r = RedirectedCommand(".zshrc", Command("cat", [], [], [], []));

julia> cmd = interpret(r)
pipeline(`cat`, stdin<Base.FileRedirect(".zshrc", false))

julia> typeof(cmd)
Base.CmdRedirect

interpret(commands::AndCommands)

Translate an AndCommands object into a Base.AndCmds that can be executed, considering the conjunction.

Examples

julia> c1 = Command("ls", [], [], [], []);

julia> c2 = Command("pwd", [], [], [], []);

julia> ac = AndCommands(c1, c2);

julia> cmd = interpret(ac)
`ls` & `pwd`

julia> typeof(cmd)
Base.AndCmds

interpret(commands::OrCommands)

Translate an OrCommands object into a Base.OrCmds that can be executed, considering the disjunction.

Examples

julia> c1 = Command("ls", [], [], [], []);

julia> c2 = Command("pwd", [], [], [], []);

julia> oc = OrCommands(c1, c2);

julia> cmd = interpret(oc)
pipeline(`ls`, stdout=`pwd`)

julia> typeof(cmd)
Base.OrCmds