API Reference

CommandComposer module

CommandComposer.Flag — Type

Flag(long_name, short_name, description)

Create a Flag object.

Arguments

long_name::String: the full name of the flag, e.g. "recursive" for a -r flag.

short_name::String: the short, usually single-letter, name of the flag, e.g. "r" for a --recursive flag.

description::String: a description of the flag for documentation purposes.

CommandComposer.Option — Type

Option(long_name, short_name, description, value)

Create an Option object.

Arguments

long_name::String: the full name of the option, e.g. "file" for a -f option.

short_name::String: the short, usually single-letter, name of the option, e.g. "f" for a --file option.

description::String: a description of the option for documentation purposes.

value::Any: the value assigned to this option.

CommandComposer.AndCommands — Type

AndCommands(a::AbstractCommand, b::AbstractCommand)

Represents a conjunction of two commands (i.e., both commands are executed).

CommandComposer.OrCommands — Type

OrCommands(a::AbstractCommand, b::AbstractCommand)

Represents a disjunction of two commands (i.e., either command is executed).

CommandComposer.Redirect — Type

Redirect(operator::String, target::String)

Represents a redirection of command output or input.

Arguments

operator::String: the operator for the redirection, can be one of ("<", "<<", ">", ">>", "2>", "&>", ">&", "2>&1").

target::String: the target file for the redirection.

CommandComposer.RedirectedCommand — Type

RedirectedCommand(command::AbstractCommand, redirect::Redirect)

Represents a command with an associated redirection.

CommandComposer.Command — Type

Command(name, flags, options, arguments, subcommands)

Represents a command to be executed, with associated flags, options, arguments, and subcommands.

Arguments

name::String: the name of the command.

flags::Vector{Flag}: any flags associated with the command.

options::Vector{Option}: any options associated with the command.

arguments::Vector{String}: any arguments to the command.

subcommands::Vector{AbstractCommand}: any subcommands to be executed in conjunction with the main command.

CommandComposer.interpret — Function

interpret(command::Command)

Translate a Command object into a Cmd that can be executed.

Examples

julia> c = Command("ls", [Flag("all", "a", "list all files")], [], [], []);

julia> interpret(c)
`ls --all`

interpret(command::RedirectedCommand)

Translate a RedirectedCommand object into a Base.CmdRedirect that can be executed, considering the redirection.

Examples

julia> c = Command("ls", [], [], [], []);

julia> r = Redirect(">", "output.txt");

julia> cr = RedirectedCommand(c, r);

julia> cmd = interpret(cr)
pipeline(`ls`, stdout>Base.FileRedirect("output.txt", false))

julia> typeof(cmd)
Base.CmdRedirect

interpret(commands::AndCommands)

Translate an AndCommands object into a Base.AndCmds that can be executed, considering the conjunction.

Examples

julia> c1 = Command("ls", [], [], [], []);

julia> c2 = Command("pwd", [], [], [], []);

julia> ac = AndCommands(c1, c2);

julia> cmd = interpret(ac)
`ls` & `pwd`

julia> typeof(cmd)
Base.AndCmds

interpret(commands::OrCommands)

Translate an OrCommands object into a Base.OrCmds that can be executed, considering the disjunction.

Examples

julia> c1 = Command("ls", [], [], [], []);

julia> c2 = Command("pwd", [], [], [], []);

julia> oc = OrCommands(c1, c2);

julia> cmd = interpret(oc)
pipeline(`ls`, stdout=`pwd`)

julia> typeof(cmd)
Base.OrCmds